swagger return value

+++ maid-care-service-backend/src/job/job.controller.ts
@@ -11,10 +11,11 @@ import {
   UnauthorizedException,
   NotFoundException,
 } from '@nestjs/common';
-import { ApiBearerAuth, ApiTags } from '@nestjs/swagger';
+import { ApiBearerAuth, ApiTags, ApiCreatedResponse } from '@nestjs/swagger';
 import { JwtAuthGuard } from '../auth/passport/jwt-auth.guard';
 import { JobService } from './job.service';
 import { CreateJobDto } from './dto/create-job.dto';
+import { JobDto } from './dto/job.dto';
 
 @Controller('job')
 @ApiTags('job')
@@ -22,28 +23,33 @@ export class JobController {
   constructor(private readonly jobService: JobService) {}
 
   @Post()
+  @ApiCreatedResponse({ type: JobDto })
   @UseGuards(JwtAuthGuard)
   @ApiBearerAuth('acess-token')
   async postJob(@Request() req, @Body() createJobDto: CreateJobDto) {
     if (req.user.role === 'customer') {
-      return await this.jobService.createJob(req.user._id, createJobDto);
+      const job = await this.jobService.createJob(req.user._id, createJobDto);
+      return new JobDto(job);
     } else throw new UnauthorizedException('user is not customer');
   }
 
   @Get(':id')
+  @ApiCreatedResponse({ type: JobDto })
   async findJob(@Param('id') id: string) {
     const job = await this.jobService.findJob(id);
     if (!job) throw new NotFoundException('job not valid');
-    return job;
+    return new JobDto(job);
   }
 
   @Delete(':id')
+  @ApiCreatedResponse({ type: JobDto })
   @UseGuards(JwtAuthGuard)
   @ApiBearerAuth('acess-token')
   async removeJob(@Request() req, @Param('id') id: string) {
     if (req.user.role === 'admin') {
       try {
-        return await this.jobService.removeJob(id);
+        const job = await this.jobService.removeJob(id);
+        return new JobDto(job);
       } catch (error) {
         throw error;
       }
@@ -51,18 +57,45 @@ export class JobController {
   }
 
   @Get('maid/:id')
+  @ApiCreatedResponse({ type: [JobDto] })
   async findByMaid(@Param('id') id: string) {
     return await this.jobService.findByMaid(id);
   }
 
   @Put(':id/reject')
+  @ApiCreatedResponse({ type: JobDto })
   @UseGuards(JwtAuthGuard)
   @ApiBearerAuth('acess-token')
   async reject(@Request() req, @Param('id') id: string) {
     const job = await this.jobService.findJob(id);
-    if (job && req.user._id == job.maidId && job.expiryTime > new Date()) {
+    if (
+      job &&
+      job.state === 'posted' &&
+      req.user._id == job.maidId &&
+      job.expiryTime > new Date()
+    ) {
       this.jobService.deleteTimeout(job);
-      return await this.jobService.reject(job);
-    } else throw new NotFoundException('job not valid');
+      await this.jobService.reject(job);
+      return new JobDto(job);
+    } else throw new NotFoundException('job not found');
+  }
+
+  @Put(':id/accept')
+  @ApiCreatedResponse({ type: JobDto })
+  @UseGuards(JwtAuthGuard)
+  @ApiBearerAuth('acess-token')
+  async accept(@Request() req, @Param('id') id: string) {
+    const job = await this.jobService.findJob(id);
+    if (
+      job &&
+      job.state === 'posted' &&
+      req.user._id == job.maidId &&
+      job.expiryTime > new Date()
+    ) {
+      this.jobService.deleteTimeout(job);
+      job.state = 'matched';
+      await job.save();
+      return new JobDto(job);
+    } else throw new NotFoundException('job not found');
   }
 }
